requirements.txt
fastapi
uvicorn
jinja2
python-multipart 
python-docx
pypdf
PyPDF2
sentence-transformers
faiss-cpu
gspread 
oauth2client